extract entries, tests passing
